pyproject.toml
[project]
name = "rag-project"
version = "0.1.0"
description = "RAG Chatbot with LangChain, Ollama, and pgvector"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "fastapi>=0.128.6",
    "uvicorn>=0.40.0",
    "python-dotenv>=1.2.1",
    "langchain>=1.2.9",
    "langchain-community>=0.4.1",
    "langchain-text-splitters>=1.1.0",
    "tortoise-orm>=0.21.8",
    "asyncpg>=0.30.0",
    "psycopg2-binary>=2.9.10",
    "pgvector>=0.3.7",
    "pypdf>=5.2.0",
    "python-docx>=1.1.2",
    "python-multipart>=0.0.20",
    "httpx>=0.28.1",
    "pydantic>=2.10.6",
    "pydantic-settings>=2.7.1",
]
